First commit: working program
Get the last 10 pulls request for nodejs organisation and node repo.
Check when was the first request approved (if there is one)
Compare with the creation date of the pull request.
Use that difference to feed an histogram data object.
Need to refactor and add github authentication to stop the limit of connection per ip